# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AiMedRes Medical AI API
  description: |
    API for AI-driven medical research and clinical decision support.
    
    **Security:** All endpoints require authentication via API key or OAuth2 token.
    **Compliance:** HIPAA/GDPR compliant with PHI de-identification and audit logging.
  version: 1.0.0
  contact:
    name: AiMedRes Support
    url: https://github.com/V1B3hR/AiMedRes
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://staging.aimedres.example.com/api/v1
    description: Staging server
  - url: https://api.aimedres.example.com/api/v1
    description: Production server
security:
  - ApiKeyAuth: []
  - OAuth2: [openid, profile, email]
...
paths:
...
  /cases:
    get:
      summary: List cases
      description: Get list of clinical cases
      operationId: listCases
      tags: [Cases]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_review, completed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Cases retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
  /cases/{case_id}:
    get:
      summary: Get case details
      description: Get detailed information for a specific case
      operationId: getCaseDetail
      tags: [Cases]
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Case details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetail'
        '404':
          description: Case not found
  /cases/{case_id}/approve:
    post:
      summary: Approve case
      description: Human-in-loop approval for case recommendation
      operationId: approveCase
      tags: [Cases]
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Case approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /auth/oidc/authorize
          tokenUrl: /auth/oidc/token
          scopes:
            openid: OpenID Connect
            profile: User profile
            email: User email
  schemas:
...
    CaseListResponse:
      type: object
      properties:
        cases:
          type: array
          items:
            $ref: '#/components/schemas/CaseSummary'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
    CaseSummary:
      type: object
      properties:
        case_id:
          type: string
        patient_id:
          type: string
        status:
          type: string
        risk_level:
          type: string
        created_at:
          type: string
          format: date-time
    CaseDetail:
      type: object
      properties:
        case_id:
          type: string
        patient_id:
          type: string
        status:
          type: string
        risk_level:
          type: string
        prediction:
          type: object
        explainability:
          type: object
        created_at:
          type: string
          format: date-time
    ApprovalRequest:
      type: object
      required: [action, rationale]
      properties:
        action:
          type: string
          enum: [approve, reject, request_review]
        rationale:
          type: string
        notes:
          type: string
    ApprovalResponse:
      type: object
      properties:
        case_id:
          type: string
        status:
          type: string
        approved_by:
          type: string
        approved_at:
          type: string
          format: date-time
...